Rearrange migrations so server doesn't need to depend on store
Puts the common code in common but keeps the actual invocation living
in the soql-server-pg
