+++ soql-server-pg/docker/ship.d/run
@@ -12,7 +12,7 @@ fi
 su socrata -c '/usr/bin/java \
     -Dconfig.file=${SERVER_CONFIG} \
     -cp $SERVER_ARTIFACT \
-    com.socrata.pg.store.Main --migrate Migrate \
+    com.socrata.pg.server.MigrateSchema Migrate \
     com.socrata.soql-server-pg.migrations \
     '
