+++ store-pg/src/main/scala/com/socrata/pg/store/Main.scala
@@ -3,9 +3,5 @@ package com.socrata.pg.store
 import com.socrata.datacoordinator.secondary.SecondaryWatcherApp
 
 object Main extends App {
-  if(args.headOption == Some("--migrate")) {
-    MigrateSchema(args.drop(1))
-  } else {
-    SecondaryWatcherApp(new PGSecondary(_))
-  }
+  SecondaryWatcherApp(new PGSecondary(_))
 }
